requirements.txt
flask>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
mistralai>=0.0.10
transformers>=4.35.0
torch>=2.0.0
ollama>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.1
pyarrow>=14.0.1
pillow>=10.0.0
soundfile>=0.12.1
webdataset>=0.2.48 